requirements.txt
alembic==1.7.7
aniso8601==9.0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
cachelib==0.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.1
email-validator==1.1.3
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.1.1
Flask-Login==0.6.0
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.29.1
gcloud==0.18.3
googleapis-common-protos==1.56.0
greenlet==1.1.2
gunicorn==20.1.0
haversine==2.5.1
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.2.0
jupyterlab-widgets==1.1.0
jws==0.1.3
kiwisolver==1.3.2
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.13
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.22.2
oauth2client==4.1.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1rc1
psutil==5.9.0
psycopg==3.0.9
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
Pygments==2.11.2
pykalman==0.9.5
pyparsing==3.0.7
Pyrebase==3.0.27
Pyrebase4==4.5.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
python-firebase==1.2.1
python-jwt==2.0.1
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2
SQLAlchemy==1.4.35
stack-data==0.2.0
terminado==0.13.3
tinycss2==1.1.1
tornado==6.1
traitlets==5.1.1
tzdata==2021.5
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.1
widgetsnbextension==3.6.0
WTForms==3.0.1
zipp==3.8.0
